changed favoriteToggle
