+++ src/components/FavoriteMovies/FavoriteMovies.module.scss
@@ -15,3 +15,10 @@
   align-self: flex-end;
   margin-right: 4em;
 }
+
+.noFavorites {
+  font-size: 1.3rem;
+  color: #888;
+  text-align: center;
+  margin-top: 20px;
+}
