+++ src/components/FavoriteToggle/FavoriteToggle.tsx
@@ -1,23 +1,39 @@
-import { FC, useState } from "react";
+import { FC, useEffect, useState } from "react";
 import { FavoriteButton } from "../FavoriteButton/FavoriteButton";
+import { useDispatch } from "react-redux";
+import {
+  addFavoriteMovie,
+  removeFavoriteMovie,
+} from "../../redux/slices/favoritesSlice";
+import { MovieDetails } from "../../types/types";
 
 export type FavoriteToggleProps = {
   isFavorite: boolean | undefined;
-  onToggle: (newIsFavorite: boolean) => void;
+  movie?: MovieDetails;
 };
 
 export const FavoriteToggle: FC<FavoriteToggleProps> = ({
   isFavorite,
-  onToggle,
+  movie,
 }) => {
+  const dispatch = useDispatch();
   const [isFavoriteState, setIsFavoriteState] = useState<boolean>(
     isFavorite ?? false
   );
 
+  useEffect(() => {
+    setIsFavoriteState(isFavorite ?? false);
+  }, [isFavorite]);
+
   const handleFavoriteToggle = () => {
     const newFavoriteState = !isFavoriteState;
     setIsFavoriteState(newFavoriteState);
-    onToggle(newFavoriteState);
+
+    if (newFavoriteState && movie) {
+      dispatch(addFavoriteMovie(movie));
+    } else if (movie) {
+      dispatch(removeFavoriteMovie({ id: movie.id }));
+    }
   };
 
   return (
